requirements.txt
# Kocaeli Üniversitesi Sınav Takvimi Sistemi
# Python Bağımlılıkları

# ============================================================
# GUI Framework - PySide6 (Qt6 - Modern ve hızlı)
# ============================================================
PySide6>=6.8.0
PySide6-Addons>=6.8.0
PySide6-Essentials>=6.8.0

# ============================================================
# Veritabanı - PostgreSQL
# ============================================================
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
asyncpg>=0.29.0  # Asenkron PostgreSQL desteği

# ============================================================
# Excel İşlemleri
# ============================================================
openpyxl>=3.1.2           # Excel okuma/yazma (XLSX)
xlrd>=2.0.1               # Eski Excel formatları (XLS)
pandas>=2.1.4             # Veri işleme ve analiz
numpy>=1.26.3             # Pandas dependency

# ============================================================
# PDF İşlemleri
# ============================================================
reportlab>=4.0.9          # PDF oluşturma
PyPDF2>=3.0.1             # PDF okuma/manipülasyon
fpdf2>=2.7.8              # Basit PDF oluşturma

# ============================================================
# Güvenlik ve Şifreleme
# ============================================================
bcrypt>=4.1.2             # Şifre hashleme
cryptography>=41.0.7      # Genel şifreleme
python-dotenv>=1.0.0      # Çevre değişkenleri

# ============================================================
# Validasyon ve Veri Kontrolü
# ============================================================
email-validator>=2.1.0    # Email validasyonu
python-dateutil>=2.8.2    # Tarih işlemleri

# ============================================================
# Optimizasyon ve Performans
# ============================================================
cachetools>=5.3.2         # Cache yönetimi
redis>=5.0.1              # Redis cache (opsiyonel)

# ============================================================
# Loglama ve Monitoring
# ============================================================
loguru>=0.7.2             # Gelişmiş loglama

# ============================================================
# Test ve Geliştirme
# ============================================================
pytest>=7.4.3             # Test framework
pytest-qt>=4.3.1          # Qt test desteği
black>=23.12.1            # Code formatter
pylint>=3.0.3             # Code analyzer

# ============================================================
# Yardımcı Kütüphaneler
# ============================================================
tqdm>=4.66.1              # Progress bar
tabulate>=0.9.0           # Tablo formatı
Pillow>=10.2.0            # Görsel işleme